Simplify command buffers and reduce coupling 💑
