+++ ecs/src/commands/mod.rs
@@ -2,17 +2,13 @@ use std::{
     alloc::Layout,
     any::{self, TypeId},
     borrow::Cow,
-    cell::RefCell,
-    rc::Rc,
 };
 
 mod command;
-mod command_buffer;
 
 pub(crate) use command::Command;
-pub(crate) use command_buffer::CommandBuffer;
 
-use crate::Entity;
+use crate::{Entity, World};
 
 #[derive(Debug, Clone, Copy)]
 pub enum EntityRef {
@@ -36,33 +32,39 @@ struct ExecutionContext {
 /// Issuing commands on a command buffer after calling `World::maintain` since creation will result
 /// in a panic.
 pub struct Commands {
-    inner: Rc<RefCell<CommandBuffer>>,
+    commands: Vec<Command>,
     entity_counter: usize,
 }
 
 impl Commands {
-    pub(crate) fn new() -> (Rc<RefCell<CommandBuffer>>, Self) {
-        let inner = Rc::new(RefCell::new(CommandBuffer::new()));
-        (
-            inner.clone(),
-            Self {
-                inner,
-                entity_counter: 0,
-            },
-        )
+    /// Retrieves a `Commands` which can be used to issue commands to be run on this `World` when
+    /// possible, which is when `maintain` is called.
+    /// After `maintain` is called, issuing more commands to the `Commands` will result in a panic.
+    pub fn new() -> Self {
+        Self {
+            commands: vec![],
+            entity_counter: 0,
+        }
+    }
+
+    /// Executes all commands on the `World` and clears the command buffer.
+    pub fn apply(&mut self, world: &mut World) {
+        let mut ctx = ExecutionContext::default();
+        for command in self.commands.drain(..) {
+            command.execute(world, &mut ctx);
+        }
+        self.entity_counter = 0;
     }
 
     pub fn spawn(&mut self) -> EntityRef {
         let id = self.entity_counter;
         self.entity_counter += 1;
-        self.inner.borrow_mut().push(Command::Spawn);
+        self.commands.push(Command::Spawn);
         EntityRef::New(id)
     }
 
     pub fn despawn<E: Into<EntityRef>>(&mut self, entity: E) {
-        self.inner
-            .borrow_mut()
-            .push(Command::Despawn(entity.into()));
+        self.commands.push(Command::Despawn(entity.into()));
     }
 
     pub fn add<T: 'static, E: Into<EntityRef>>(&mut self, entity: E, component: T) {
@@ -72,7 +74,7 @@ impl Commands {
         let component = Box::new(component);
         let component = Box::into_raw(component);
         let component = component as *mut u8;
-        self.inner.borrow_mut().push(Command::AddComponent {
+        self.commands.push(Command::AddComponent {
             entity: entity.into(),
             type_id: TypeId::of::<T>(),
             name: Cow::Borrowed(any::type_name::<T>()),
